Fixed warning: comparison between signed and unsigned integer expressions
git-svn-id: http://svn.code.sf.net/p/protomol/svn/trunk@1982 ce46b6ef-3445-0410-911d-c47028e6f9df

+++ protomol/src/protomol/force/hessian/ReducedHessGB.cpp
@@ -269,7 +269,7 @@ Real ReducedHessGB::FirstDerivativeFGB(
 
   for(unsigned k=0; k<sz; k++){
     
-    if( k == atom1 ) continue;
+    if( k == (unsigned)atom1 ) continue;
       
     //grab r_{ik} / r_{jk}
     const Real dist = topo->atoms[atom1].myGBSA_T->distij[k];
